Create Base_URL =
This command specifies the initial position of Iteration, it indicates the total number(s) of spoken language(s) to be used to infoData Chesscryption, it also indicates the direction of Iterative move(s). This command is the most important script of the ACS-Web-X Program. Base_URL is the latest communicational identity, you can call that a big Digital Asset...